Delay indexing by a second so that everything has time to settle
Indexing immediately when the event arrives sometimes causes issues
trying to stat the changed resource.
